Refactor FurnaceMind chat UI helpers
